Sorted out some transitions. Battery screen shows for 2 seconds then switches back
